+++ src/test/java/com/sparta/easyspring/commentlike/controller/CommentLikeControllerTest.java
@@ -7,8 +7,8 @@ import com.sparta.easyspring.auth.security.UserDetailsServiceImpl;
 import com.sparta.easyspring.auth.util.JwtUtil;
 import com.sparta.easyspring.commentlike.repository.CommentLikeRepository;
 import com.sparta.easyspring.commentlike.service.CommentLikeService;
-import com.sparta.easyspring.postlike.config.MockTestDataSetup;
-import com.sparta.easyspring.postlike.controller.MockSpringSecurityFilter;
+import com.sparta.easyspring.config.MockSpringSecurityFilter;
+import com.sparta.easyspring.config.MockTestDataSetup;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
