+++ src/test/java/com/sparta/easyspring/postlike/controller/PostLikeControllerTest.java
@@ -1,12 +1,12 @@
 package com.sparta.easyspring.postlike.controller;
 
-import com.sparta.easyspring.MockSpringSecurityFilter;
+import com.sparta.easyspring.config.MockSpringSecurityFilter;
 import com.sparta.easyspring.auth.config.SecurityConfig;
 import com.sparta.easyspring.auth.entity.User;
 import com.sparta.easyspring.auth.security.UserDetailsImpl;
 import com.sparta.easyspring.auth.security.UserDetailsServiceImpl;
 import com.sparta.easyspring.auth.util.JwtUtil;
-import com.sparta.easyspring.postlike.config.MockTestDataSetup;
+import com.sparta.easyspring.config.MockTestDataSetup;
 import com.sparta.easyspring.postlike.repository.PostLikeRepository;
 import com.sparta.easyspring.postlike.service.PostLikeService;
 import org.junit.jupiter.api.*;
